Suppress subsequent errors in svg_check.sh
This code was introduced via I02715a5. In case there is something
wrong with the svgo call (e.g. when my local node_modules are not
up to date), I get many unrelated error messages that are more
confusing than helpful. The first error on the svgo call is still
shown.

Bug: T280923
Change-Id: I5ef7843c9de122c8dce61fbb98737c8acdd0c22c