Add relative positioning to .mwe-popups-container
The Popups UI currently depends on the absolute positioning of the
top-level .mwe-popups element to correctly position it's footer.

This dependence on absolute positioning makes it difficult to debug
Popups in isolation (e.g. in Storybook).

Adding relative positioning to the .mwe-popups-container elements
maintains the correct footer positioning and removes the need for a
specialized layout in Storybook that accommodates absolute positioning.

(If changing the Popups layout to accommodate Storybook seems
unjustified, just think of it as increasing testability).

Bug: T271763
Change-Id: Ifba23afa1a2f401b025476a1fa464c2176fb40ec