Fixes Selenium Popups tests
As part of T310054, we moved article tabs into
the #content area on the beta cluster which caused
the daily tests to fail. As a result the page previews
test is currently expecting a page preview to show
on hovering "Article" which is not the case.

Update selector to target links in the parser
content

Bug: T310313
Change-Id: I0bbccbbfbae458d5c32eaea12f855b832b52491c