Record canonical special page names in virtual pageviews
When analyzing event data special pages are a bit unique. They
are the same across all wikis, but the name is typically
localized to the wiki itself. The localization is relatively
constant, but can change in ways that the canonical special page
names do not. Simplify analyzing virtualpageviews from special
pages by recording the canonical name of the page, rather than
the localized name.

Changes the VirtualPageViews event submission to provide
wgCanonicalSpecialPageName when appropriate.

Bug: T375387
Change-Id: Ia344d8d60058082fdf0715281d732e5c1f90a38a