Kick bundlesize out of package.json
This pulled in ancient versions of brotli-size (unlikely to be fixed:
https://github.com/siddharthkp/bundlesize/pull/370#issuecomment-1038968845)
which broke CI on Node 20. IIUC, it shouldn’t even be needed anymore now
that bundle size can be tested via BundleSizeTestBase instead; but in
any case this can be figured out later, *after* CI is restored for the
many extensions that pull in Popups.

Bug: T382192
Bug: T360590
Change-Id: I86413a7707cbd33fd473f8a1daf6c7866fbb473f