Clean up bundlesize config
This follows up the parent change, where I think I misunderstood the
bundlesize configs. The configs in bundlesize.config.json and
package.json are separate, and the latter became unused by kicking out
npm bundlesize; but I think that can be addressed by merging it into the
former. The size has to be a bit larger to pass CI – I assume because it
now includes styles, messages, and types.json as well.

Bug: T382192
Bug: T360590
Change-Id: I91dcf50e3ea7d4dcb68189b63a1a0d60368dccf9