styles: Use `calc()` with Codex token `@size-icon-medium`
This token will soon point to a CSS custom property. To use CSS
custom properties in mathematical operations, you must use the
CSS `calc()` function.

Note that in my testing, even though `size-icon-medium` is in rems
and `width-icon-gutter` is in ems, the resulting number is as
expected and the resulting value is in rems.

Bug: T389116
Change-Id: If89f9e72c912a7d00a7393853a9b9d551b6d08ed