ui/renderer: fallback settings URL for anons
Logged-out users have no SpecialPage for settings. So PreviewBehavior
provides an undefined value for `settingsUrl` in this case.
Previously, it was casted to string "undefined": while hovering the cog
button, the user might think it linked to `/wiki/undefined`. This patch
use hash '#' as fallback for `settingsUrl`, so, while hovering the cog
button, the user may think they will stay on current page if they click
on it.

Change-Id: I6ed255be346603d1814e243a1444487af9171007