Respect wgThumbnailSteps when generating thumbs
Popups was generating a lot of non-standard thumb sizes to fit the
little window; 320px, 480px, and 640px as well as other sizes for
tall aspect ratios.

Have added a common JS-side function to implement the logic around
following wgThumbnailSteps to match File's behavior, and calling it
here.

In the long run it would be wise to move more of the box-fitting
logic to mw.util as well, as many extensions seem to implement
it separately.

Bug: T411013
Depends-On: I9a9a8b888a69505cb5d1babd7e1606761f68456c
Change-Id: I0227cd31ecd6b1bf9d4c89f1ebcb2691fec28582